fix critical json-importer bug
Fixes #1544.

+++ plugin/RunalyzePluginTool_DbBackup/class.RunalyzeJsonImporter.php
@@ -97,7 +97,7 @@ class RunalyzeJsonImporter {
 	 */
 	private function deleteOldData() {
 		$Requests = array(
-			'delete_trainings'	=> array('training', 'route', 'trackdata', 'swimdata', 'hrv', 'activity_equipment'),
+			'delete_trainings'	=> array('training', 'route'),
 			'delete_user_data'	=> array('user'),
 		);
 
@@ -115,7 +115,7 @@ class RunalyzeJsonImporter {
 	 * @param string $table without prefix
 	 */
 	private function truncateTable($table) {
-		$this->Results->addDeletes('runalyze_'.$table, $this->DB->query('DELETE FROM `'.PREFIX.$table.'`')->rowCount());
+		$this->Results->addDeletes('runalyze_'.$table, $this->DB->query('DELETE FROM `'.PREFIX.$table.'` WHERE `accountid`="'.$this->AccountID.'"')->rowCount());
 	}
 
 	/**
@@ -314,7 +314,14 @@ class RunalyzeJsonImporter {
 			$Row = current($CompleteRow);
 			$Values = array_values($Row);
 
-			if ($Columns[0] == 'name' || $TableName == 'runalyze_plugin') {
+			if (in_array($TableName, array(
+				'runalyze_equipment',
+				'runalyze_equipment_type',
+				'runalyze_plugin',
+				'runalyze_route',
+				'runalyze_sport',
+				'runalyze_type'
+			))) {
 				if (isset($this->ExistingData[$TableName][$Values[0]])) {
 					$this->ReplaceIDs[$TableName][$ID] = $this->ExistingData[$TableName][$Values[0]];
 				} else {
